+++ pkg/database/models/rating.go
@@ -22,9 +22,9 @@ type RatingEntry struct {
 	Queue        string `gorm:"type:queue_type"`
 	Tier         string `gorm:"type:tier_type"`
 	Rank         string `gorm:"type:rank_type"`
-	LeaguePoints uint16
-	Wins         uint16
-	Losses       uint16
+	LeaguePoints int
+	Wins         int
+	Losses       int
 	Region       regions.SubRegion `gorm:"type:varchar(5)"`
 	FetchTime    time.Time         `gorm:"autoCreateTime"`
 }
